webservice/requirements.txt
# Web Service specific requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Core ML dependencies (shared with other components)
mlflow==2.8.1
scikit-learn==1.3.2
xgboost==2.0.2
pandas==2.1.4
numpy==1.25.2

# AWS/S3 dependencies
boto3==1.34.0
python-dotenv==1.0.0

# Additional utilities
python-multipart==0.0.6
requests==2.31.0
